A lot of fixes and plumbing
