+++ packages/dev/core/src/Engines/engineStore.ts
@@ -1,7 +1,7 @@
 import { Observable } from "../Misc/observable";
 import type { Nullable } from "../types";
 
-import type { Engine } from "./engine";
+import type { AbstractEngine } from "./abstractEngine";
 import type { Scene } from "../scene";
 
 /**
@@ -10,13 +10,13 @@ import type { Scene } from "../scene";
  */
 export class EngineStore {
     /** Gets the list of created engines */
-    public static Instances: Engine[] = [];
+    public static Instances: AbstractEngine[] = [];
 
     /**
      * Notifies when an engine was disposed.
      * Mainly used for static/cache cleanup
      */
-    public static OnEnginesDisposedObservable = new Observable<Engine>();
+    public static OnEnginesDisposedObservable = new Observable<AbstractEngine>();
 
     /** @internal */
     public static _LastCreatedScene: Nullable<Scene> = null;
@@ -24,7 +24,7 @@ export class EngineStore {
     /**
      * Gets the latest created engine
      */
-    public static get LastCreatedEngine(): Nullable<Engine> {
+    public static get LastCreatedEngine(): Nullable<AbstractEngine> {
         if (this.Instances.length === 0) {
             return null;
         }
