+++ packages/tools/testTools/src/visualizationUtils.ts
@@ -1,4 +1,7 @@
 /* eslint-disable no-console */
+
+import type { Engine } from "core/Engines/engine";
+
 /* eslint-disable @typescript-eslint/naming-convention */
 declare const BABYLON: typeof window.BABYLON;
 
@@ -244,7 +247,7 @@ export const evaluateDisposeSceneForVisualization = async (engineFlags: { forceU
 
 export const evaluateIsGLError = async () => {
     try {
-        const gl = window.engine!._gl,
+        const gl = (window.engine! as Engine)._gl,
             glError = gl ? gl.getError() : 0;
         if (gl && glError !== 0) {
             return true;
