+++ packages/tools/tests/test/visualization/webgpu_webgpu_log.txt
@@ -1,0 +1,82 @@
+preparing page
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF %cDownload the React DevTools for a better development experience: https://reactjs.org/link/react-devtools font-weight:bold
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF %cDownload the React DevTools for a better development experience: https://reactjs.org/link/react-devtools font-weight:bold
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF %cDownload the React DevTools for a better development experience: https://reactjs.org/link/react-devtools font-weight:bold
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF %cDownload the React DevTools for a better development experience: https://reactjs.org/link/react-devtools font-weight:bold
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF [webpack-dev-server] Server started: Hot Module Replacement disabled, Live Reloading disabled, Progress disabled, Overlay enabled.
+INF %cDownload the React DevTools for a better development experience: https://reactjs.org/link/react-devtools font-weight:bold
+page ready
+LOG BJS - [09:22:53]: Babylon.js v7.0.0 - WebGPU1 engine
+engine ready
+Running - edge-renderer-and-zOffset
+LOG BJS - [09:22:55]: Babylon.js v7.0.0 - WebGPU1 engine
+engine ready
+Running - cube-with-holes-using-stencil-buffer
+LOG BJS - [09:22:57]: Babylon.js v7.0.0 - WebGPU1 engine
+engine ready
+Running - water-material
+this.createCubeTextureBase is not a function
+ERR JSHandle@error
+LOG BJS - [09:23:00]: Babylon.js v7.0.0 - WebGPU1 engine
+engine ready
+Running - show-bounding-box
+LOG BJS - [09:23:01]: Babylon.js v7.0.0 - WebGPU1 engine
+engine ready
+Running - cube-reflection-with-probe-and-mirrors
+this._engine.createDepthStencilTexture is not a function
+ERR JSHandle@error
+LOG BJS - [09:23:04]: Babylon.js v7.0.0 - WebGPU1 engine
+engine ready
+Running - sphere-with-custom-shader-to-display-wireframe-using-glow-layer
+this.createCubeTextureBase is not a function
+ERR JSHandle@error
+LOG BJS - [09:23:07]: Babylon.js v7.0.0 - WebGPU1 engine
+engine ready
+Running - blur-cube-with-the-effect-renderer
+this._engine.createDepthStencilTexture is not a function
+ERR JSHandle@error
+LOG BJS - [09:23:10]: Babylon.js v7.0.0 - WebGPU1 engine
+engine ready
+Running - simple-render-target-with-blue-spheres
+this._engine.createDepthStencilTexture is not a function
+ERR JSHandle@error
+LOG BJS - [09:23:13]: Babylon.js v7.0.0 - WebGPU1 engine
+engine ready
+Running - neon-pipe-with-glow-layer
+ERR BJS - [09:23:16]: Unable to load from https://www.babylonjs.com/Assets/NeonPipe/glTF/NeonPipe.gltf: Error in onSuccess callback
+Cannot read properties of undefined (reading 'renderPassId')
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
+INF [webpack-dev-server] Disconnected!
+INF [webpack-dev-server] Trying to reconnect...
