+++ packages/tools/viewer/src/labs/texture.ts
@@ -4,6 +4,7 @@ import { InternalTexture, InternalTextureSource } from "core/Materials/Textures/
 import { Scalar } from "core/Maths/math.scalar";
 import type { BaseTexture } from "core/Materials/Textures/baseTexture";
 import { Texture } from "core/Materials/Textures/texture";
+import type { Engine } from "core/Engines/engine";
 
 /**
  * WebGL Pixel Formats
@@ -251,7 +252,7 @@ export class TextureUtils {
 
                 const gl = (<any>scene.getEngine())._gl;
                 gl.bindTexture(gl.TEXTURE_CUBE_MAP, glTexture);
-                scene.getEngine()._unpackFlipY(false);
+                (scene.getEngine() as Engine)._unpackFlipY(false);
                 if (face instanceof HTMLElement || face instanceof ImageData) {
                     gl.texImage2D(gl.TEXTURE_CUBE_MAP_POSITIVE_X + i, level, textureCube.internalFormat, textureCube.internalFormat, textureCube.type, <any>face);
                 } else {
